src/requirements.txt
zipfile38==0.0.3
seaborn==0.11.0
simpledbf==0.2.6
pandas==1.1.3
numpy==1.19.2
ipython==7.24.1
python_dateutil==2.8.1
scikit_learn==0.24.2
jupyter
